Manage MT-32 threaded rendering and playback using the moodycamel ring buffer
The ring buffer makes the:
 - mixer-callback wait until a rendered buffer is available before playback
 - rendering thread wait until space is available in the ring before rendering

Previously, playback position relative to rendering position along with thread
state management was needed inside the mixer callback to prevent the rendering
thread from running away.  Now, with rendering managed by the ring buffer, we
can eliminate:
 - mutex, locking, and conditional coordination sent from the mixer
 - current vs previous per-frame render-position management
 - the above occured every millisecond (on callback), which also goes away

We also can eliminate the equivalent management code in the rendering thread,
which was needed to coordinate the rendering position with playback position.

Rendering is now also performed in fixed-sized large chunks, which is more
efficient than performing using smaller and/or per-millisecond chunks.

The ring buffer also performs std::move() operations of the buffers to ensure
no copies are created.